Update README.md
Turn in the CODE in D2L and PDF report in Gradescope.

Your code, especially Part 2, will be tested. Make sure it follows the template posted on d2l (assignment section) for naming your functions and parameters each function takes. If your code is missing, has errors, or doesn’t follow the template, it will be marked as “failed to be tested” and points will be deducted without the opportunity for reconsideration.

Ensure that your code is well-commented to facilitate understanding of the methods used to generate your results.

The full names of all members should be included in the first page of the report as well as name of the file (it can include only last names).

The team size must be between 2-4 members.

Each team is required to submit their report and code only once per team, and the same team member should submit on both D2L and GradeScope platforms. Multiple submission by same person is allowed.